+++ CHANGELOG.md
@@ -6,3 +6,4 @@
  - add readme file
  - add documentation folder
  - add installation in readme file
+ - accept limited list of combinators